[MOV] website: move social media related tests to tests/builder
The tests in "social_media_and_share.test.js" file were located in the
interactions folder, although they do not test any interaction related
behavior. Since social media related tests already exist in
"social_media.test.js" file under the tests/builder folder, these tests
are moved there for better consistency and organization.

Part of this commit: https://github.com/odoo/odoo/commit/a7489557f04ab2d617996ed56e7647c5fb7cbed1
